requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
django==2.2.6
djangorestframework==3.10.3
pytz==2019.3              # via django
sqlparse==0.3.0           # via django
